refactor: extract make_cli_cmd into agentmux-common shared crate
Moves the .cmd wrapper parser into agentmux-common so both agentmux-srv
and agentmux-cef use a single implementation. Also syncs
package-lock.json to 0.33.69.
