+++ agentmux-cef/Cargo.toml
@@ -20,6 +20,7 @@ sandbox = ["cef/sandbox"]
 resources_path = "resources"
 
 [dependencies]
+agentmux-common = { path = "../agentmux-common" }
 cef = { version = "146", default-features = false, features = ["build-util"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
