+++ agentmux-common/src/lib.rs
@@ -1,0 +1,5 @@
+//! Shared utilities for AgentMux crates.
+
+mod cli;
+
+pub use cli::make_cli_cmd;
